+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - PIM-9668: Asset attribute - Fix Text area + Rich text editor modes turn text attributes into infinite extendable fields when not using spaces
 - PIM-9664: Display Ziggy as asset image when the preview cannot be generated
 - PIM-9702: Fix infinite loop when using get all assets API endpoint
+- PIM-9720: Fix Asset cursor search after query
 
 ## Improvements